Fix memory handling bug
The CIDToGID map is an array of ints.  The code properly allocated
a number N of ints, but then used memcpy for N unsigned shorts.
That left the upper half of the array uninitialized.

+++ qt5/src/ArthurOutputDev.cc
@@ -673,7 +673,7 @@ void ArthurOutputDev::updateFont(GfxState *state)
         if (codeToGIDLen) {
           codeToGID = (int *)gmallocn(codeToGIDLen, sizeof(int));
           memcpy(codeToGID, ((GfxCIDFont *)gfxFont)->getCIDToGID(),
-                 codeToGIDLen * sizeof(unsigned short));
+                 codeToGIDLen * sizeof(int));
         }
       } else {
         auto ff = (fontLoc->locType != gfxFontLocEmbedded)
